src/test/java/codewars/com/minions/CountCharactersTest.java
package codewars.com.minions;

import org.junit.Test;

import static org.junit.Assert.assertEquals;

/**
 * Unit Tests for CountCharacters.
 */
public class CountCharactersTest {

    /**
     * Sample test provided by CodeWars.
     */
    @Test
    public void testCodeWars() {
        final int expectedFour = 4;
        final int expectedFive = 5;

        assertEquals(expectedFour, CountCharacters.countChar("fizzbuzz", 'z'));
        assertEquals(expectedFive, CountCharacters.countChar("Fancy fifth fly aloof", 'f'));
        assertEquals(expectedFour, CountCharacters.countChar("fizzbuzz", 'Z'));
        assertEquals(expectedFive, CountCharacters.countChar("Fancy fifth fly aloof", 'F'));
    }
}
